Move TCP server into sub-module
